fix(ui): address departure time review feedback
