+++ app/ui/src/tests/vehicleDepartureTime.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from "vitest";
 
-import { getMinutesAfterHourChange } from "@/app/edit/components/vehicle/VehicleDetailsOverlay";
+import { getMinutesAfterHourChange } from "@/app/edit/utils/deliveryHelpers";
 
 describe("getMinutesAfterHourChange", () => {
   it("defaults empty minutes to 00 when the hour is valid", () => {
@@ -11,7 +11,6 @@ describe("getMinutesAfterHourChange", () => {
   it("keeps minutes that the user already entered", () => {
     expect(getMinutesAfterHourChange("9", "15")).toBe("15");
     expect(getMinutesAfterHourChange("12", "30")).toBe("30");
-    expect(getMinutesAfterHourChange("13", "00", false)).toBe("00");
   });
 
   it("leaves minutes empty when the hour is missing or invalid", () => {
@@ -26,5 +25,9 @@ describe("getMinutesAfterHourChange", () => {
 
   it("clears auto-filled minutes when the hour becomes invalid", () => {
     expect(getMinutesAfterHourChange("13", "00", true)).toBe("");
+  });
+
+  it("preserves user-entered 00 when the hour becomes invalid", () => {
+    expect(getMinutesAfterHourChange("13", "00", false)).toBe("00");
   });
 });
